Refactor getRequiredMerge
