+++ __tests__/unit/common/blockchain/testPlasma.js
@@ -1,6 +1,5 @@
 import {
   Plasma,
-  Utxos,
   Ethereum,
   TxDetails,
   BlockchainParams
@@ -15,10 +14,9 @@ jest.mock('common/blockchain/gasEstimator.js')
 jest.mock('common/blockchain/wait.js')
 jest.mock('common/blockchain/ethereum.js')
 jest.mock('common/blockchain/txDetails.js')
-jest.mock('@omisego/react-native-omg-js')
 jest.mock('common/blockchain/contract.js')
 
-const { getBalance, getUtxos } = PlasmaClient.ChildChain
+const { getBalance } = PlasmaClient.ChildChain
 const { getDeposit } = TxDetails
 const { signSendTx } = Ethereum
 const { TEST_ADDRESS, TEST_PRIVATE_KEY } = Config
@@ -27,10 +25,6 @@ const FIVE_GWEI = '5000000000'
 
 const mockGetBalancesResponse = resp => {
   getBalance.mockReturnValueOnce(Promise.resolve(resp))
-}
-
-const mockGetUtxosResponse = resp => {
-  getUtxos.mockReturnValueOnce(Promise.resolve(resp))
 }
 
 const mockGetDepositTxDetails = resp => {
@@ -57,386 +51,6 @@ describe('Test Plasma Boundary', () => {
           amount: '5000000000',
           currency: '0x0000000000000000000000000000000000000000'
         }
-      ])
-    })
-  })
-
-  test('getUtxos should return BN amount as a string', () => {
-    const utxos = [
-      {
-        amount: new BN('5000000000000000000'),
-        blknum: 1010,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1010000000000
-      }
-    ]
-    mockGetUtxosResponse(utxos)
-
-    return Utxos.get(TEST_ADDRESS).then(data => {
-      expect(data).toEqual([
-        {
-          amount: '5000000000000000000',
-          blknum: 1010,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1010000000000
-        }
-      ])
-    })
-  })
-
-  test('getUtxos should return utxos where `utxo.currency` === `currency`', () => {
-    const utxos = [
-      {
-        amount: new BN('5000000000000000000'),
-        blknum: 1010,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1010000000000
-      },
-      {
-        amount: new BN('10000000000000000'),
-        blknum: 1006,
-        currency: '0x0000000000000000000000000000000000000000',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1006000000000
-      }
-    ]
-    mockGetUtxosResponse(utxos)
-
-    const requestOptions = {
-      currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5'
-    }
-
-    return Utxos.get(TEST_ADDRESS, requestOptions).then(data => {
-      expect(data).toEqual([
-        {
-          amount: '5000000000000000000',
-          blknum: 1010,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1010000000000
-        }
-      ])
-    })
-  })
-
-  test('getUtxos should return all utxos if not given `currency`', () => {
-    const utxos = [
-      {
-        amount: new BN('5000000000000000000'),
-        blknum: 1010,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1010000000000
-      },
-      {
-        amount: new BN('10000000000000000'),
-        blknum: 1006,
-        currency: '0x0000000000000000000000000000000000000000',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1006000000000
-      }
-    ]
-    mockGetUtxosResponse(utxos)
-
-    const requestOptions = {}
-
-    return Utxos.get(TEST_ADDRESS, requestOptions).then(data => {
-      expect(data).toEqual([
-        {
-          amount: '5000000000000000000',
-          blknum: 1010,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1010000000000
-        },
-        {
-          amount: '10000000000000000',
-          blknum: 1006,
-          currency: '0x0000000000000000000000000000000000000000',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1006000000000
-        }
-      ])
-    })
-  })
-
-  test('getUtxos should return utxos where `utxo.currency` === `currency` and `utxo.utxo_pos` >= `fromUtxoPos`', () => {
-    const utxos = [
-      {
-        amount: new BN('5000000000000000000'),
-        blknum: 1010,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1010000000000
-      },
-      {
-        amount: new BN('1000000000000000000'),
-        blknum: 1008,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1008000000000
-      },
-      {
-        amount: new BN('1000000000000000000'),
-        blknum: 1007,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1007000000000
-      },
-      {
-        amount: new BN('10000000000000000'),
-        blknum: 1006,
-        currency: '0x0000000000000000000000000000000000000000',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1006000000000
-      }
-    ]
-    mockGetUtxosResponse(utxos)
-
-    const requestOptions = {
-      currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-      fromUtxoPos: 1008000000000
-    }
-
-    return Utxos.get(TEST_ADDRESS, requestOptions).then(data => {
-      expect(data).toEqual([
-        {
-          amount: '5000000000000000000',
-          blknum: 1010,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1010000000000
-        },
-        {
-          amount: '1000000000000000000',
-          blknum: 1008,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1008000000000
-        }
-      ])
-    })
-  })
-
-  test('getUtxos should return utxos where `utxo.utxo_pos` >= `fromUtxoPos`', () => {
-    const utxos = [
-      {
-        amount: new BN('5000000000000000000'),
-        blknum: 1010,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1010000000000
-      },
-      {
-        amount: new BN('1000000000000000000'),
-        blknum: 1008,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1008000000000
-      },
-      {
-        amount: new BN('10000000000000000'),
-        blknum: 1006,
-        currency: '0x0000000000000000000000000000000000000000',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1006000000000
-      },
-      {
-        amount: new BN('1000000000000000000'),
-        blknum: 1007,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1005000000000
-      }
-    ]
-    mockGetUtxosResponse(utxos)
-
-    const requestOptions = {
-      fromUtxoPos: 1006000000000
-    }
-    return Utxos.get(TEST_ADDRESS, requestOptions).then(data => {
-      expect(data).toEqual([
-        {
-          amount: '5000000000000000000',
-          blknum: 1010,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1010000000000
-        },
-        {
-          amount: '1000000000000000000',
-          blknum: 1008,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1008000000000
-        },
-        {
-          amount: '10000000000000000',
-          blknum: 1006,
-          currency: '0x0000000000000000000000000000000000000000',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1006000000000
-        }
-      ])
-    })
-  })
-
-  test('getUtxos should sort utxos by `utxo_pos` in descending', () => {
-    const utxos = [
-      {
-        amount: new BN('10000000000000000'),
-        blknum: 1006,
-        currency: '0x0000000000000000000000000000000000000000',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1010000000000
-      },
-      {
-        amount: new BN('1000000000000000000'),
-        blknum: 1008,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1020000000000
-      },
-      {
-        amount: new BN('5000000000000000000'),
-        blknum: 1010,
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        oindex: 0,
-        owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-        txindex: 0,
-        utxo_pos: 1030000000000
-      }
-    ]
-    mockGetUtxosResponse(utxos)
-
-    const requestOptions = {}
-    return Utxos.get(TEST_ADDRESS, requestOptions).then(data => {
-      expect(data).toEqual([
-        {
-          amount: '5000000000000000000',
-          blknum: 1010,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1030000000000
-        },
-        {
-          amount: '1000000000000000000',
-          blknum: 1008,
-          currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1020000000000
-        },
-        {
-          amount: '10000000000000000',
-          blknum: 1006,
-          currency: '0x0000000000000000000000000000000000000000',
-          oindex: 0,
-          owner: '0x4522fb44c2ab359e76ecc75c22c9409690f12241',
-          txindex: 0,
-          utxo_pos: 1010000000000
-        }
-      ])
-    })
-  })
-
-  test('getRequiredMerge should returns utxos with length > 1', () => {
-    const utxos = [
-      {
-        amount: '10000000000000000',
-        currency: '0x0000000000000000000000000000000000000000',
-        utxo_pos: 1,
-        owner: TEST_ADDRESS
-      },
-      {
-        amount: '1000000000000000000',
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        utxo_pos: 2,
-        owner: TEST_ADDRESS
-      },
-      {
-        amount: '5000000000000000000',
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        utxo_pos: 3,
-        owner: TEST_ADDRESS
-      },
-      {
-        amount: '2000000000000000000',
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        utxo_pos: 4,
-        owner: TEST_ADDRESS
-      },
-      {
-        amount: '3000000000000000000',
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        utxo_pos: 5,
-        owner: TEST_ADDRESS
-      },
-      {
-        amount: '7000000000000000000',
-        currency: '0xa1c9d0c6ed627fb2197fd16cd3afde37cc5e8da5',
-        utxo_pos: 6,
-        owner: TEST_ADDRESS
-      }
-    ]
-    mockGetUtxosResponse(utxos)
-    return Utxos.getRequiredMerge(TEST_ADDRESS, null, 4).then(arrayOfUtxos => {
-      expect(arrayOfUtxos).toEqual([
-        [utxos[5], utxos[4], utxos[3], utxos[2], utxos[1]]
       ])
     })
   })
